Refactor state management examples and setup for Redux Toolkit, TanStack Query, NgRx, Pinia and Zustand Moved example code and data to separate data files for better organization and maintainability.
