+++ src/pages/StateManagement/NgRx/Ngrxsetup.jsx
@@ -3,51 +3,13 @@ import { SiNgrx } from "react-icons/si";
 import { useState } from "react";
 import { HiCheckCircle } from "react-icons/hi";
 import { IoCopyOutline } from "react-icons/io5";
-
-const steps = [
-  {
-    title: "Install @ngrx packages",
-    command: "npm install @ngrx/store @ngrx/effects @ngrx/entity @ngrx/store-devtools",
-    description: "Core NgRx modules for store, side effects, entity, and developer tools.",
-  },
-  {
-    title: "Register Store in AppModule",
-    command: `// app.module.ts
-import { StoreModule } from '@ngrx/store';
-import { effects } from './state';
-@NgModule({
-  imports: [
-    StoreModule.forRoot({}),
-    EffectsModule.forRoot(effects),
-  ],
-})`,
-    description: "Initialize NgRx store and load root effects.",
-  },
-  {
-    title: "Add Store DevTools",
-    command: `// app.module.ts
-import { StoreDevtoolsModule } from '@ngrx/store-devtools';
-StoreDevtoolsModule.instrument({ maxAge: 25, logOnly: environment.production })`,
-    description: "Enable time-travel debugging and state inspection.",
-  },
-];
-
-const resources = [
-  { name: "NgRx Docs", link: "https://ngrx.io/guide/store", description: "Official guide and API reference." },
-  { name: "NgRx GitHub", link: "https://github.com/ngrx/platform", description: "Source code and issues." },
-  { name: "NgRx Demo App", link: "https://github.com/ngrx/platform/tree/master/projects/example-app", description: "Work through a full demo application." },
-];
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
+import { handleCopy } from "@/components/shared/CopyToClipboard"
+import { packageManagers, steps, resources } from "@/pages/StateManagement/NgRx/NgRxData"
 
 export const NgrxSetup = () => {
   const [copied, setCopied] = useState(null);
-
-  const handleCopy = async (cmd, idx) => {
-    try {
-      await navigator.clipboard.writeText(cmd);
-      setCopied(idx);
-      setTimeout(() => setCopied(null), 1500);
-    } catch { }
-  };
+  const [selectedTabs, setSelectedTabs] = useState({});
 
   return (
     <div className="space-y-12 px-4 md:px-6 py-10">
@@ -81,8 +43,15 @@ export const NgrxSetup = () => {
               </div>
               <div className="relative">
                 <button
-                  onClick={() => handleCopy(step.command, idx)}
-                  className="text-muted-foreground hover:text-primary transition"
+                  onClick={() => handleCopy(
+                    typeof step.command === "string"
+                      ? step.command
+                      : step.command[selectedTabs[idx] || "npm"], // pick per-step tab
+                    setCopied,
+                    idx
+                  )}
+                  className="text-muted-foreground hover:text-[#319795] transition"
+                  title="Copy to clipboard"
                 >
                   {copied === idx ? (
                     <HiCheckCircle className="w-6 h-6 text-green-600" />
@@ -97,9 +66,35 @@ export const NgrxSetup = () => {
                 )}
               </div>
             </div>
-            <pre className="bg-muted/40 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+            {/* <pre className="bg-muted/40 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
               {step.command}
-            </pre>
+            </pre> */}
+            {step.all ? (
+              <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                {step.command}
+              </pre>
+            ) : (
+              <Tabs value={selectedTabs[idx] || "npm"} onValueChange={(value) => setSelectedTabs((prev) => ({ ...prev, [idx]: value }))} className="w-full">
+                <TabsList className="">
+                  {packageManagers.map((pm) => (
+                    <TabsTrigger key={pm.key} value={pm.key}>
+                      <span className="flex items-center gap-1 text-xs cursor-pointer">
+                        {pm.icon}
+                        {pm.label}
+                      </span>
+                    </TabsTrigger>
+                  ))}
+                </TabsList>
+
+                {packageManagers.map((pm) => (
+                  <TabsContent key={pm.key} value={pm.key}>
+                    <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                      {step.command[pm.key]}
+                    </pre>
+                  </TabsContent>
+                ))}
+              </Tabs>
+            )}
           </div>
         ))}
       </div>
