+++ src/pages/StateManagement/ReduxToolkit/ReduxSetup.jsx
@@ -1,81 +1,15 @@
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { TerminalSquare } from "lucide-react";
 import { useState } from "react";
-import { FaReact } from "react-icons/fa";
 import { HiCheckCircle } from "react-icons/hi";
 import { IoCopyOutline } from "react-icons/io5";
 import { SiRedux } from "react-icons/si";
-
-const steps = [
-    {
-        title: "Install Redux Toolkit and React-Redux",
-        command: "npm install @reduxjs/toolkit react-redux",
-        description: "Core dependencies for store configuration and React bindings.",
-    },
-    {
-        title: "Create the Store",
-        command: `// store.js
-import { configureStore } from '@reduxjs/toolkit';
-import counterReducer from './features/counterSlice';
-
-export const store = configureStore({
-  reducer: {
-    counter: counterReducer,
-  },
-});`,
-        description: "Use `configureStore` to bootstrap your Redux store with slices.",
-    },
-    {
-        title: "Wrap the App with Provider",
-        command: `// main.jsx or index.js
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import { Provider } from 'react-redux';
-import { store } from './store';
-
-ReactDOM.createRoot(document.getElementById('root')).render(
-  <Provider store={store}>
-    <App />
-  </Provider>
-);`,
-        description: "Integrate Redux with React using the `Provider` component.",
-    },
-];
-const resources = [
-    {
-        name: "Redux Toolkit Docs",
-        link: "https://redux-toolkit.js.org",
-        description: "The official Redux Toolkit documentation, including tutorials and API reference.",
-    },
-    {
-        name: "Redux Toolkit GitHub",
-        link: "https://github.com/reduxjs/redux-toolkit",
-        description: "GitHub repository for RTK—open issues, source code, discussions.",
-    },
-    {
-        name: "Redux Essentials Tutorial",
-        link: "https://redux.js.org/tutorials/essentials/part-1-overview-concepts",
-        description: "Step-by-step tutorial using Redux Toolkit in real-world apps.",
-    },
-    {
-        name: "RTK Query Docs",
-        link: "https://redux-toolkit.js.org/rtk-query/overview",
-        description: "Built-in solution for data fetching and caching in Redux Toolkit.",
-    },
-];
+import { handleCopy } from "@/components/shared/CopyToClipboard"
+import { packageManagers, steps, resources} from "@/pages/StateManagement/ReduxToolkit/ReduxData"
 
 export const ReduxSetup = () => {
     const [copied, setCopied] = useState(null);
-
-    const handleCopy = async (text, idx) => {
-        try {
-            await navigator.clipboard.writeText(text);
-            setCopied(idx);
-            setTimeout(() => setCopied(null), 1500);
-        } catch (err) {
-            console.error("Copy failed", err);
-        }
-    };
+    const [selectedTabs, setSelectedTabs] = useState({});
 
     return (
         <div className="space-y-12 px-4 md:px-6 py-10">
@@ -108,8 +42,15 @@ export const ReduxSetup = () => {
                             </div>
                             <div className="relative">
                                 <button
-                                    onClick={() => handleCopy(step.command, idx)}
-                                    className="text-muted-foreground hover:text-primary transition"
+                                    onClick={() => handleCopy(
+                                        typeof step.command === "string"
+                                            ? step.command
+                                            : step.command[selectedTabs[idx] || "npm"], // pick per-step tab
+                                        setCopied,
+                                        idx
+                                    )}
+                                    className="text-muted-foreground hover:text-[#319795] transition"
+                                    title="Copy to clipboard"
                                 >
                                     {copied === idx ? (
                                         <HiCheckCircle className="w-6 h-6 text-green-600" />
@@ -124,9 +65,35 @@ export const ReduxSetup = () => {
                                 )}
                             </div>
                         </div>
-                        <pre className="bg-muted/50 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                        {/* <pre className="bg-muted/50 border border-[#cccccc] rounded-md p-3 mt-2 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
                             {step.command}
-                        </pre>
+                        </pre> */}
+                        {step.all ? (
+                            <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                                {step.command}
+                            </pre>
+                        ) : (
+                            <Tabs value={selectedTabs[idx] || "npm"} onValueChange={(value) => setSelectedTabs((prev) => ({ ...prev, [idx]: value }))} className="w-full">
+                                <TabsList className="">
+                                    {packageManagers.map((pm) => (
+                                        <TabsTrigger key={pm.key} value={pm.key}>
+                                            <span className="flex items-center gap-1 text-xs cursor-pointer">
+                                                {pm.icon}
+                                                {pm.label}
+                                            </span>
+                                        </TabsTrigger>
+                                    ))}
+                                </TabsList>
+
+                                {packageManagers.map((pm) => (
+                                    <TabsContent key={pm.key} value={pm.key}>
+                                        <pre className="bg-muted/50 border border-gray-300 rounded-md p-3 font-mono text-sm whitespace-pre-wrap overflow-x-auto">
+                                            {step.command[pm.key]}
+                                        </pre>
+                                    </TabsContent>
+                                ))}
+                            </Tabs>
+                        )}
                     </div>
                 ))}
             </div>
